backend/pyproject.toml
[project]
name = "med-track-backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.8"
dependencies = [
    "fastapi>=0.124.4",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
    "requests>=2.32.4",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.33.0",
]
